feat: pick video image + edit video schema + view realtime video status

+++ lib/features/video_render/data/model/video_render_model.dart
@@ -12,7 +12,7 @@ class VideoRenderModel extends VideoRender {
     return VideoRenderModel(
       id: map['id'] as String,
       status: map['status'] as String,
-      progress: map['progress'] as int,
+      progress: (map['progress'] as double).toInt(),
       createdAt: DateTime.parse(map['created_at']),
       updatedAt: DateTime.parse(map['updated_at']),
     );
